+++ lib/presentation/match_setting/match_setting_model.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class MatchSettingModel extends ChangeNotifier {
+  final MatchNameController = TextEditingController();
+  final ATeamNameController = TextEditingController();
+  final BTeamNameController = TextEditingController();
+  final ServiceController = TextEditingController();
+  var OutText = '';
+  List FileContentsList = ['', '', '', ''];
+  var items = ['ATeam', 'BTeam'];
+  var firstServe;
+}
