+++ lib/presentation/point_counting/point_counting_model.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+class PointCountingModel extends ChangeNotifier {}
