+++ pubspec.yaml
@@ -22,6 +22,7 @@ dependencies:
     sdk: flutter
   path_provider: ^1.6.9
   flutter_share: ^1.0.2+1
+  provider: ^4.1.3+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
